pyproject.toml
[tool.poetry]
name = "quotient-bot"
version = "3.5.0"
description = ""
authors = ["deadaf <askpygod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.2"
tortoise-orm = "^0.17.2"
colorama = "^0.4.4"
asyncpg = "^0.22.0"
dateparser = "^1.0.0"
discord-ext-menus = "^1.1"
aerich = "^0.5.3"
jishaku = "^1.20.0"
parsedatetime = "^2.6"
Pillow = "^8.2.0"
"mystbin.py" = "^2.1.3"
prettytable = "^2.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.5b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 122